requirements.txt
absl-py==2.1.0
aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.9.1
aiosignal==1.3.1
alchemy-sdk==0.1.1
alembic==1.14.1
amqp==5.3.1
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
appnope==0.1.4
APScheduler==3.11.0
arch==7.2.0
argon2-cffi==23.1.0
arrow==1.3.0
asn1crypto==1.5.1
asttokens==3.0.0
astunparse==1.6.3
async-generator==1.10
async-lru==2.0.5
async-timeout==4.0.3
atomicwrites==1.4.0
attrs==24.2.0
audioread==3.0.1
Authlib==1.6.3
babel==2.17.0
backoff==2.2.1
base58==2.1.1
bcrypt==4.2.0
bidict==0.23.1
billiard==4.2.1
bip-utils==2.9.3
bitcoin==1.1.42
bitpay==6.1.0
bleach==6.1.0
blinker==1.8.2
blockfrost-python==0.6.0
bokeh==3.5.2
boto==2.49.0
boto3==1.35.32
botocore==1.35.32
cached-property==1.5.2
cachelib==0.9.0
cardano==0.8.2
cardano-tools==2.1.0
cbor2==5.6.4
cdp==0.0.2
cdp-sdk==0.11.0
celery==5.3.0
certifi==2024.8.30
certvalidator==0.11.1
cffi==1.17.1
ckzg==2.0.1
clear==2.0.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cmudict==1.0.32
coinbase==2.1.0
coinbase-advanced-py==1.6.2
coincurve==20.0.0
coloredlogs==15.0.1
comm==0.2.2
common==0.1.2
conan==2.8.0
contourpy==1.3.0
cose==0.9.dev8
crcmod==1.7
cryptoapis==1.8.0
cryptography==43.0.1
curl_cffi==0.12.0
cycler==0.12.1
cytoolz==0.12.3
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.14
decorator==5.2.1
Deprecated==1.2.18
dnslib==0.9.25
dnspython==2.7.0
docker==7.1.0
dogecoin-python==0.1.3
ecdsa==0.19.0
ECPy==1.2.5
ed25519-blake2b==1.4.1
email_validator==2.2.0
et_xmlfile==2.0.0
eth-account==0.13.3
eth-bloom==2.0.0
eth-keys==0.5.1
eth-pydantic-types==0.1.3
eth-rlp==2.1.0
eth-tester==0.9.1b1
eth-typing==5.0.0
eth-utils==5.0.0
eventlet==0.36.1
evm-trace==0.1.3
executing==2.2.0
fastapi==0.115.5
fastecdsa==1.7.5
fasteners==0.19
fastjsonschema==2.21.1
Flask-Bcrypt==1.0.1
Flask-Caching==2.3.0
flask-cors==5.0.1
Flask-Executor==1.0.0
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-PyMongo==3.0.1
flask-restx==1.3.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flatbuffers==24.3.25
fonttools==4.53.1
fqdn==1.5.1
frozendict==2.4.4
frozenlist==1.4.1
futureproof==0.3.1
gast==0.6.0
gevent==24.11.1
google==3.0.0
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.1.1
grpcio==1.68.1
gunicorn==23.0.0
h11==0.16.0
h2==4.1.0
hexbytes==1.2.1
hpack==4.0.0
httpcore==1.0.9
httplib2==0.30.0
httpx==0.28.1
humanfriendly==10.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jwt==1.3.1
kaggle==1.7.4.2
keras==3.7.0
kiwisolver==1.4.7
kombu==5.4.2
libclang==18.1.1
libp2p==0.2.0
librosa==0.11.0
limits==5.1.0
lxml==5.3.0
Mako==1.3.9
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
ml-dtypes==0.4.1
mnemonic==0.21
mpld3==0.5.10
msgspec==0.18.6
multiaddr==0.0.9
multidict==6.1.0
multitasking==0.0.11
mypy-protobuf==3.6.0
Naked==0.1.32
namex==0.0.8
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
netaddr==1.3.0
ngrok==1.4.0
noiseprotocol==0.3.1
notebook==7.4.1
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.3.1
ogmios==1.2.1
openai==0.11.5
opencv-python==4.10.0.84
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.13.1
ordered-set==4.1.0
orjson==3.10.7
oscrypto==1.3.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandas-stubs==2.2.3.241009
pandocfilters==1.5.1
paramiko==3.5.0
parso==0.8.4
patch-ng==1.18.0
patsy==1.0.1
pbr==6.1.1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.2
plotly==5.24.1
pooch==1.8.2
pprintpp==0.4.0
priority==2.0.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
pronouncing==0.2.0
propcache==0.2.0
proto-plus==1.26.1
protobuf==6.30.2
psutil==7.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.9
pure_eval==0.2.3
py-algorand-sdk==2.6.1
py-ecc==6.0.0
py-evm==0.7.0a4
py-geth==5.0.0
py-solc-x==2.0.3
py-sr25519-bindings==0.2.0
pyasn1-modules==0.2.8
pybind11==2.13.6
pycardano==0.12.0
pychrome==0.2.4
pycoin==0.92.20230326
pycparser==2.22
pycryptodome==3.23.0
pycryptodomex==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
pyethash==0.1.27
PyGithub==2.7.0
Pygments==2.19.1
PyJWT==2.9.0
pymongo==4.11.1
pymultihash==0.8.2
PyNaCl==1.5.0
pyparsing==3.1.4
PyQt5==5.15.10
PyQtWebEngine==5.15.6
pyroute2==0.7.12
python-bitcoinlib==0.12.2
python-dogecoin==0.0.4
python-dotenv==1.1.0
python-engineio==4.11.2
python-json-logger==3.3.0
python-pptx==1.0.2
python-slugify==8.0.4
python-socketio==5.12.1
PyYAML==6.0.2
pyzmq==26.4.0
qiskit==2.0.0
qiskit-aer==0.17.0
qiskit-algorithms==0.3.1
Quart==0.19.9
redis==4.6.0
referencing==0.35.1
regex==2024.9.11
requests-futures==1.0.2
requests-oauthlib==2.0.0
RestrictedPython==8.0
retrying==1.3.4
rich==13.9.4
rlp==4.0.1
rpcudp==5.0.0
rpds-py==0.21.0
rq==1.16.2
rsa==4.9.1
rsconnect_python==1.24.0
rustworkx==0.16.0
s3transfer==0.10.2
safe-pysha3==1.0.4
scandir==1.10.0
schedule==1.2.2
scikit-learn==1.5.2
scipy==1.12.0
semantic-version==2.10.0
semver==2.13.0
Send2Trash==1.8.3
sentry-sdk==2.14.0
setuptools==75.6.0
shellescape==3.8.1
sigfig==1.3.3
simple-websocket==1.1.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
SQLAlchemy==2.0.32
stack-data==0.6.3
starlette==0.41.3
statsmodels==0.14.4
stevedore==5.4.1
stripe==10.10.0
symengine==0.13.0
sympy==1.14.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
terminado==0.18.1
theblockchainapi==0.0.604
threadpoolctl==3.5.0
tinycss2==1.4.0
toolz==0.12.1
torch==2.6.0
tqdm==4.67.1
tqdm-loggable==0.2
trie==2.2.0
trio==0.26.2
trio-typing==0.10.0
twilio==9.4.6
typeguard==4.3.0
types-Deprecated==1.2.9.20240311
types-protobuf==5.28.0.20240924
types-python-dateutil==2.9.0.20241206
types-pytz==2024.2.0.20241003
types-requests==2.32.0.20240914
types-urllib3==1.26.25.14
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.3
u-msgpack-python==2.8.0

uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.2.3
uvicorn==0.31.0
varint==1.0.2
vine==5.1.0
wcwidth==0.2.13
web3==7.2.0
web3-ethereum-defi==0.26
webcolors==24.11.1
websocket==0.2.1
websockets==13.0.1
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.17.2
wsproto==1.2.0
WTForms==3.2.1
wxPython==4.2.2
XlsxWriter==3.2.0
xrpl-py==3.0.0
xyzservices==2024.9.0
yahooquery==2.3.7
yarl==1.11.1
yfinance==0.2.65
zipp==3.21.0
zope.event==5.0
zope.interface==7.1.1
zstandard==0.23.0
